make components global by dax
